Don't handle GestureTap events in WebViewPlugin
If we forward the GestureTap event to WebView, it will be converted to mouse events. And the click event will open a link if there is one.
However, the parent frame has no idea about this. So it will convert the gestureTap event again and pass the mouse events to WebViewPlugin. Thus causing a link to be clicked twice.

BUG=162140


Review URL: https://chromiumcodereview.appspot.com/12226064

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181434 0039d316-1c4b-4281-b951-d872f2087c98