Reland 181499 - Allow Terms of Service to be specified for Public Accounts
This CL allows Terms of Service to be specified for a Public Account. The
terms are set by specifying a URL through policy. Chrome downloads the
terms and presents them to the user. Only if the user accepts the terms
can the Public Account be entered.

This CL also fixes the regression described in issue 174314.

TEST=Manual
BUG=chromium-os:32780, 174314
TBR=nkostylev,mnissler

Review URL: https://codereview.chromium.org/12221097

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181656 0039d316-1c4b-4281-b951-d872f2087c98