Make devtools_window.cc multi-desktop aware.
Getting rid of BrowserList:: methods which only know about the native-desktop.

Note that CreateBrowserForDevTools is still native desktop only so that devtools will not fully work yet in a multi-desktop environment, but this is making progress in the refactoring goal of getting rid of all native desktop only methods.

BUG=129187


Review URL: https://chromiumcodereview.appspot.com/12210063

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181697 0039d316-1c4b-4281-b951-d872f2087c98