[Telemetry] Disconnect a Tab's inspector backend when it's GC'd.
Tabs are stored in a weakref.WeakValueDictionary so that they disconnect their websocket when not in use. There's no guarantee, however, that its inspector_backend is GC'd at this time, so the websocket may stay connected. We need to explicitly disconnect it on GC.

We rely on the disconnect to happen immediately, since Chrome devtools doesn't return websocketDebuggerUrls for connected tabs, so we would have no way of reconnecting to that websocket.

BUG=166243
TEST=None.

Review URL: https://codereview.chromium.org/12241002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181750 0039d316-1c4b-4281-b951-d872f2087c98