Smooth scrolling effect to mosaic view.
In slide view we have nice sliding transition effect, while in mosaic view not. This patch adds a simple sliding animation, which is especially visible when using a keyboard. Along the way, adds scrolling margins, so it is possible to navigate without using keyboard. Refer to a screenshot in the bug description.

TEST=Open a directory (on Drive) with lots of pictures, enter mosaic view, scroll using keyboard.
BUG=174575

Review URL: https://codereview.chromium.org/12208042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181860 0039d316-1c4b-4281-b951-d872f2087c98