Fix missing callback chaining in UsbService.
On OS_CHROMEOS when the permission_broker checks a set of devices the
UsbService needs to wait for the async work to complete before responding.
This worked on OS_CHROMEOS desktop builds because the stub implementation of
PermissionBrokerClient wasn't racing to modify the devices vector.

BUG=chromium-os:38836

Review URL: https://chromiumcodereview.appspot.com/12226137

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182320 0039d316-1c4b-4281-b951-d872f2087c98