cc: Only allow trees created at the current viewport size to draw.
When we're on a small page whose WebKit layout size is equal to the
viewport size, and the viewport size grows, we don't have enough content
to fill the screen until the pending tree finishes rasterizing.  This
patch disallows drawing on the active tree when the viewport size
has recently changed.

NOTRY=true
BUG=173449


Review URL: https://chromiumcodereview.appspot.com/12212156

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182357 0039d316-1c4b-4281-b951-d872f2087c98