fix third_party_headers to work properly on Windows w/ ninja.
The setup_third_party.py script was hand-rolling an implementation of
os.path.relpath in order to be 2.5 compatible, and that was getting
confused when we had to generate forwarding headers from ninja's
intermediate directory.

This patch reworks the python code to use relpath (since we require 2.6 now)
and work the same way regardless of platform or headers being generated, so
that we can have a single code path.

Full details in the bug.

R=tony@chromium.org
BUG=175273

Review URL: https://codereview.chromium.org/12224087

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182359 0039d316-1c4b-4281-b951-d872f2087c98