History: Fix selecting multiple visits.
Initially the visist IDs for the checkboxes were set as the visits were
retrieved but currently the visits can be displayed in a different order.

Fix this by setting the ID when the visit is displayed, fixing multiple
selection. Add a test to check that in the future.

BUG=170690

Review URL: https://chromiumcodereview.appspot.com/12217015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182470 0039d316-1c4b-4281-b951-d872f2087c98