Dismiss context menu on mousewheel not over menu for Win Aura
1. The non-Aura version was lying in the comment, this code
is required for non-"Thinkpad mousewheel driver" too.

2. In non-Aura, m_hWnd is generally != GetForegroundWindow(): m_hWnd
is the content area, and GetForegroundWindow is the top level frame.
So, despite appearances this EnumThreadWindows runs on every mousewheel.
(It's possible there's some edge cases like mousewheeling in long
bookmarks or something where the != is necessary?)

3. So, do the same thing on Aura.

I'm not sure that the behaviour is really correct on non-Aura either, TBH.
Explorer and Notepad for example, just eats mousewheels (sends them to
the menu I guess) rather than dismissing the menu. But that's Windows only
so it could be we want consistency with other platforms for Aura.

In any case, this makes Aura and non-Aura match for now.

R=sky@chromium.org
BUG=174760

Review URL: https://chromiumcodereview.appspot.com/12262031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182560 0039d316-1c4b-4281-b951-d872f2087c98