webui: Allow AutocompleteList to customize the behavior for Enter key
By default, pressing the Enter key clears and hides the autocomplete popup.
This change makes the behavior customizable, so clients of AutocompleteList
can set a different behavior, like opening the selected web page.

Along the way, rename syncWidthToInput() to syncWidthAndPositionToInput()
and let it change the position, in addition to the width, as the position
of the input field may be changed, when the window is resized.
For instance, this  happens when the window is maximized.

BUG=174584
TEST=Go to "Settings" -> "Appearance" -> Enable "Show Home button" -> "Change". Confirm that the Home page selection works as before. While the suggestions are shown, maximize the window and confirm that the suggestions are layed out correctly.

Review URL: https://codereview.chromium.org/12221169

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182595 0039d316-1c4b-4281-b951-d872f2087c98