Register all components for each policy domain at once.
The PolicyService keeps track of (domain, component) pairs of policy namespaces,
so that the policy providers know which namespaces have policy consumers. This
change makes the registration of namespaces a bulk operation per policy domain,
which avoids the re-registration issue and makes each registration a full
snapshot of that domain (this allows flushing caches on startup, for example).

BUG=163318


Review URL: https://chromiumcodereview.appspot.com/12254038

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182738 0039d316-1c4b-4281-b951-d872f2087c98