Make TestWebIDBFactory allocate a separate WebIDBFactory per thread.
This allows the main WebKit thread and web worker threads to each get their own
WebIDBFactory instance, which mimics the behavior when we let the WebIDBFactory
be instantiated via the WebKitPlatformSupport::idbFactory() method.

Avoid storing WebString on class that will be reached by multiple threads.

This is a slight variation on the patch attempted here:
https://chromiumcodereview.appspot.com/12181010

R=alecflett@chromium.org

Review URL: https://codereview.chromium.org/12230054

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182828 0039d316-1c4b-4281-b951-d872f2087c98