Add "Offline" root in the left nav of Files App.
The files view for "Offline" is implemented as a search.
Changing current directory to the fake "Offline" invokes a DirectoryContentsDriveOffline search.
The specialized search traverses the tree in "My Drive" and shows the files available offline.
All done in JS and the logic is encapsulated within DirectoryModel and DirectoryContents.

BUG=168742
TEST=Open "Offline" in Files App and verify that the files available offline are listed.


Review URL: https://chromiumcodereview.appspot.com/12217005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@183017 0039d316-1c4b-4281-b951-d872f2087c98