Fix crash on Android when the Main activity is resumed after it was destroyed.
NetworkChangeNotifierAutoDetect was storing and using a pointer to the main
activity's Context that can be stale in certain circumstances.
This is problematic in case the Main activity is destroyed and then recreated
which happens for example when the user changes the system language in the
Android settings.

BUG=176292

Review URL: https://codereview.chromium.org/12294021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@183217 0039d316-1c4b-4281-b951-d872f2087c98