WebPopups not dismissed when clicking on other window
On ChromeOS Web Popups were not dismissed for mouse events
outside of the Popup or its Parent. For instance, clicking
on the desktop backgound or in the browser caption area
(perhaps for dragging). The reason is that we don't lose
focus or deactive for those events. Solution: added event
filter to look for these events and then dismiss the popup
explicitly.

BUG=175543
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/12253040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@183645 0039d316-1c4b-4281-b951-d872f2087c98