CrOS: Schedule renderer drawing to hardware vsync.
Before this patch, the renderer scheduled drawing assuming a 60Hz refresh rate
at an arbitrary timebase. With this patch, we send a hardware vsync signal to
the renderer. The timbase is offset slightly so that the browser UI and
renderer do not draw at the same time.

BUG=157223


Review URL: https://chromiumcodereview.appspot.com/12218055

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@183696 0039d316-1c4b-4281-b951-d872f2087c98