RendererWebKitPlatformSupportImpl::createAudioDevice() needs the ability to make proper audio input device selection at the hardware level.  WebKit passes us this information via an |input_device_id|
For the moment, the audio back-end is only able to support the "default" input device with
synchronized audio I/O.  So we validate that when WebKit requests audio input, that it is indeed the
"default" input device.  If not, then we ignore the request for audio input.
Review URL: https://codereview.chromium.org/12218106

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@184161 0039d316-1c4b-4281-b951-d872f2087c98