Remove the WebContentsModalDialog interface
This CL is part 6 of 6 in the work to remove the WebContentsModalDialog
interface.  Now that all references to WebContentsModalDialog are
gone, we can remove the class and the inheritance of the class by the
platform-specific UI code.

This builds on part 5 in https://codereview.chromium.org/12283024. 

BUG=157161


Review URL: https://chromiumcodereview.appspot.com/12288018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@184589 0039d316-1c4b-4281-b951-d872f2087c98