Mac: Fix dictionary popup failing for some text at end of line.
Make the selected text comparison strip any trailing/leading
whitespace. This happens to be the case for the selected text
for some end-of-line text selections.

Adds a unit test and also documents the selection-related
member variables in render_view_impl.h and fixes a few
lint warnings in the files.

BUG=170091
TEST=Per bug repro steps.

Review URL: https://chromiumcodereview.appspot.com/12295037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@184823 0039d316-1c4b-4281-b951-d872f2087c98