Initialize BrowserPolicyConnector after the IOThread is created.
The import process on Windows spawns a subprocess and spins a nested message loop waiting for it to complete. If this subprocess takes longer than 5 seconds then the message loop will execute a delayed initialization task of the policy code, that grabs the system request context; however, since the IOThread hasn't been created yet this leads to a crash.

BUG=177843

Review URL: https://chromiumcodereview.appspot.com/12310100

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@184967 0039d316-1c4b-4281-b951-d872f2087c98