cc: Remove Layer::didUpdateBounds
This function was only needed to get tests passing.  Tests are updated to
correctly call calculateDrawProperties where needed so that content bounds are
up to date.

The hud is now created in updateLayers (so that its content bounds can be
updated) rather than in willCommit.

R=danakj@chromium.org
BUG=166023


Review URL: https://chromiumcodereview.appspot.com/11743033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@185063 0039d316-1c4b-4281-b951-d872f2087c98